move the dylib/grpc client to external_storage
They don't need to be in external_storage_export
