+++ components/external_storage/Cargo.toml
@@ -7,24 +7,39 @@ publish = false
 [features]
 default = ["protobuf-codec"]
 protobuf-codec = [
-  "kvproto/protobuf-codec",
   "tikv_util/protobuf-codec",
+  "grpcio/protobuf-codec",
 ]
 prost-codec = [
-  "kvproto/prost-codec",
   "tikv_util/prost-codec",
+  "grpcio/prost-codec",
+]
+cloud-storage-dylib = [
+  "ffi-support",
+  "libloading",
+  "protobuf",
+  "tokio-util",
+]
+cloud-storage-grpc = [
+  "grpcio",
+  "tokio-util",
 ]
 
 [dependencies]
+anyhow = "1.0"
 bytes = "0.5.3"
+ffi-support = { optional = true, version = "0.4.2" }
 file_system = { path = "../file_system" }
 futures = "0.3"
 futures-executor = "0.3"
 futures-io = "0.3"
 futures-util = { version = "0.3", default-features = false, features = ["io"] }
+grpcio = { optional = true, version = "0.8",  default-features = false, features = ["openssl-vendored"] }
 kvproto = { git = "https://github.com/pingcap/kvproto.git", default-features = false }
 lazy_static = "1.3"
+libloading = { optional = true, version = "0.7.0" }
 prometheus = { version = "0.10", default-features = false, features = ["nightly", "push"] }
+protobuf = { optional = true, version = "2" }
 rand = "0.7"
 rusoto_core = "0.45.0"
 slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debug"] }
@@ -34,6 +49,7 @@ tame-oauth = "0.4.2"
 tikv_alloc = { path = "../tikv_alloc" }
 tikv_util = { path = "../tikv_util", default-features = false }
 tokio = { version = "0.2.13", features = ["time"] }
+tokio-util = { optional = true, version = "0.3.1", features = ["compat"] }
 url = "2.0"
 
 [dev-dependencies]
