+++ components/external_storage/export/src/grpc_service.rs
@@ -1,114 +1,15 @@
 // Copyright 2021 TiKV Project Authors. Licensed under Apache-2.0.
-use external_storage::ExternalStorage;
 
 use std::io::{self, ErrorKind};
 use std::sync::Arc;
 
-use crate::request::{
-    anyhow_to_io_log_error, file_name_for_write, restore_receiver, restore_sender, write_receiver,
-    write_sender, DropPath,
-};
+use crate::request::{restore_receiver, write_receiver};
 use anyhow::Context;
-use futures_io::AsyncRead;
+use external_storage::request::anyhow_to_io_log_error;
 use grpcio::{self};
 use kvproto::backup as proto;
-#[cfg(feature = "prost-codec")]
-pub use kvproto::backup::storage_backend::Backend;
-#[cfg(feature = "protobuf-codec")]
-pub use kvproto::backup::StorageBackend_oneof_backend as Backend;
 use slog_global::{error, info};
-use tikv_util::time::Limiter;
 use tokio::runtime::{Builder, Runtime};
-
-struct ExternalStorageClient {
-    backend: Backend,
-    runtime: Arc<Runtime>,
-    rpc: proto::ExternalStorageClient,
-    name: &'static str,
-    url: url::Url,
-}
-
-pub fn new_client(
-    backend: Backend,
-    name: &'static str,
-    url: url::Url,
-) -> io::Result<Box<dyn ExternalStorage>> {
-    let runtime = Builder::new()
-        .basic_scheduler()
-        .thread_name("external-storage-grpc-client")
-        .core_threads(1)
-        .enable_all()
-        .build()?;
-    Ok(Box::new(ExternalStorageClient {
-        backend,
-        runtime: Arc::new(runtime),
-        rpc: new_rpc_client()?,
-        name,
-        url,
-    }))
-}
-
-impl ExternalStorage for ExternalStorageClient {
-    fn name(&self) -> &'static str {
-        self.name
-    }
-
-    fn url(&self) -> io::Result<url::Url> {
-        Ok(self.url.clone())
-    }
-
-    fn write(
-        &self,
-        name: &str,
-        reader: Box<dyn AsyncRead + Send + Unpin>,
-        content_length: u64,
-    ) -> io::Result<()> {
-        info!("external storage writing");
-        (|| -> anyhow::Result<()> {
-            let file_path = file_name_for_write(&self.name, &name);
-            let req = write_sender(
-                &self.runtime,
-                self.backend.clone(),
-                file_path.clone(),
-                name,
-                reader,
-                content_length,
-            )?;
-            info!("grpc write request");
-            self.rpc
-                .write(&req)
-                .map_err(rpc_error_to_io)
-                .context("rpc write")?;
-            info!("grpc write request finished");
-            DropPath(file_path);
-            Ok(())
-        })()
-        .context("external storage write")
-        .map_err(anyhow_to_io_log_error)
-    }
-
-    fn read(&self, _name: &str) -> Box<dyn AsyncRead + Unpin> {
-        unimplemented!("use restore instead of read")
-    }
-
-    fn restore(
-        &self,
-        storage_name: &str,
-        restore_name: std::path::PathBuf,
-        expected_length: u64,
-        speed_limiter: &Limiter,
-    ) -> io::Result<()> {
-        info!("external storage restore");
-        let req = restore_sender(
-            self.backend.clone(),
-            storage_name,
-            restore_name,
-            expected_length,
-            speed_limiter,
-        )?;
-        self.rpc.restore(&req).map_err(rpc_error_to_io).map(|_| ())
-    }
-}
 
 #[derive(Debug)]
 pub struct SocketService {
@@ -137,14 +38,6 @@ pub fn new_service() -> io::Result<SocketService> {
     })()
     .context("new service")
     .map_err(anyhow_to_io_log_error)
-}
-
-fn new_rpc_client() -> io::Result<proto::ExternalStorageClient> {
-    let env = Arc::new(grpcio::EnvBuilder::new().build());
-    let grpc_socket_path = "/tmp/grpc-external-storage.sock";
-    let socket_addr = format!("unix:{}", grpc_socket_path);
-    let channel = grpcio::ChannelBuilder::new(env).connect(&socket_addr);
-    Ok(proto::ExternalStorageClient::new(channel))
 }
 
 /// Service handles the RPC messages for the `ExternalStorage` service.
@@ -226,21 +119,6 @@ pub fn make_rpc_error(err: io::Error) -> grpcio::RpcStatus {
         },
         Some(format!("{:?}", err)),
     )
-}
-
-pub fn rpc_error_to_io(err: grpcio::Error) -> io::Error {
-    let msg = format!("{}", err);
-    match err {
-        grpcio::Error::RpcFailure(status) => match status.status {
-            grpcio::RpcStatusCode::NOT_FOUND => io::Error::new(ErrorKind::NotFound, msg),
-            grpcio::RpcStatusCode::INVALID_ARGUMENT => io::Error::new(ErrorKind::InvalidInput, msg),
-            grpcio::RpcStatusCode::UNAUTHENTICATED => {
-                io::Error::new(ErrorKind::PermissionDenied, msg)
-            }
-            _ => io::Error::new(ErrorKind::Other, msg),
-        },
-        _ => io::Error::new(ErrorKind::Other, msg),
-    }
 }
 
 fn bind_socket(
