+++ components/external_storage/export/src/lib.rs
@@ -4,9 +4,9 @@ mod export;
 pub use export::*;
 
 #[cfg(feature = "cloud-storage-grpc")]
-mod service;
+mod grpc_service;
 #[cfg(feature = "cloud-storage-grpc")]
-pub use service::new_service;
+pub use grpc_service::new_service;
 
 #[cfg(feature = "cloud-storage-dylib")]
 mod dylib;
