+++ components/external_storage/src/lib.rs
@@ -26,6 +26,12 @@ mod noop;
 pub use noop::NoopStorage;
 mod metrics;
 use metrics::EXT_STORAGE_CREATE_HISTOGRAM;
+#[cfg(feature = "cloud-storage-dylib")]
+pub mod dylib_client;
+#[cfg(feature = "cloud-storage-grpc")]
+pub mod grpc_client;
+#[cfg(any(feature = "cloud-storage-dylib", feature = "cloud-storage-grpc"))]
+pub mod request;
 
 pub fn record_storage_create(start: Instant, storage: &dyn ExternalStorage) {
     EXT_STORAGE_CREATE_HISTOGRAM
